requirements-gpu.txt
numpy
scipy
sklearn
matplotlib
pandas
keras
tensorflow-gpu==1.4.0
h5py